feat(ui): refactor task form
